just so much shit
moved to using browserify for building the file
