+++ index.html
@@ -17,7 +17,7 @@
     <link rel="stylesheet" href="style.css" />
 
     <!-- import the webpage's javascript file -->
-    <script src="script.js" defer></script>
+    <script src="dist/bundle.js" defer></script>
   </head>
   <body></body>
 </html>
