+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "gophers",
+  "version": "1.0.0",
+  "description": "a gopher game overlay for OBS, for [rosco](https://www.twitch.tv/roscomcc)",
+  "main": "src/main.js",
+  "scripts": {
+    "start": "concurrently \"npm run server\" \"npm run build:watch\"",
+    "server": "serve .",
+    "build:watch": "watchify src/main.js -o dist/bundle.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/zaccolley/gophers.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/zaccolley/gophers/issues"
+  },
+  "homepage": "https://github.com/zaccolley/gophers#readme",
+  "devDependencies": {
+    "concurrently": "^5.3.0",
+    "serve": "^11.3.2",
+    "watchify": "^3.11.1"
+  }
+}
