+++ style.css
@@ -5,7 +5,7 @@
 }
 
 body {
-  color: white;
+  font-family: sans-serif;
   background-color: rgba(0, 0, 0, 0);
   overflow: hidden;
   height: 100vh;
@@ -13,7 +13,7 @@ body {
 }
 
 button {
-  width: 100vh;
+  width: 100vw;
   height: 100vh;
   font-size: 5em;
 }
